PHP validation forms, css improved, functionality added

+++ Project_WEB/configurations/bookingsconfig/deleteBooking.php
@@ -1,11 +1,14 @@
 <?php 
+if (!isset($_SESSION)) {
+    session_start();
+  }
     include '../bookingsMapper.php';
+    include '../../views/Dashboard/message.php';
     $mapper = new BookingsMapper();
     $id = $_REQUEST['booking_ID'];
     $delete = $mapper->deleteBooking($id);
  
     if ($delete) {
         header("Location:../../views/Dashboard/BookingsManagement.php");
-        echo "<script>alert('Booking deleted successfully');</script>";
     }
  ?>
